expose getScale
