+++ package.json
@@ -20,16 +20,16 @@
     "test": "builder run npm:test"
   },
   "dependencies": {
-    "builder": "~2.1.4",
-    "builder-victory-component": "~0.0.5",
-    "d3-scale": "^0.3.0",
+    "builder": "~2.4.0",
+    "builder-victory-component": "~0.0.9",
     "victory-animation": "^0.0.13",
     "victory-label": "^0.1.8",
-    "victory-util": "^2.0.3"
+    "victory-util": "file:../victory-util"
   },
   "devDependencies": {
-    "builder-victory-component-dev": "~0.0.5",
+    "builder-victory-component-dev": "~0.0.9",
     "chai": "^3.2.0",
+    "d3-scale": "^0.3.0",
     "mocha": "^2.3.3",
     "react": "^0.14.0",
     "react-addons-test-utils": "^0.14.0",
